SynthesiaSet/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
include = ["src/**/*.py"]

[tool.black]
line-length = 120

[project]
name = "synthesiaset"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    # Utils + Config
    "black==25.9.0",
    "tyro==0.9.35",
    "matplotlib==3.10.7",
    "pillow==12.0.0",
    "tqdm==4.67.1",
    # Core
    "numpy==2.3.4",
    "mitsuba==3.7.1",
    "drjit==1.2.0",
    "trimesh==4.9.0",
    "manifold3d==3.2.1",
]
authors = [
    {name = "Matan Davidi"},
    {name = "Cyril Moser"},
    {name = "Gent Serifi", email = "gent.serifi@gmail.com"},
    {name = "Nicola Studer", email = "nicola.studer2000@gmail.com"},
]
description = "Synthetic Piano Dataset Generator"
readme = "README.md"

[project.scripts]
synthesia = "main:entrypoint"